restructuring the code layout
