+++ prr/services/providers/anthropic/complete.py
@@ -5,9 +5,9 @@ import os
 
 import anthropic
 
-from prr.config import load_config
-from prr.request import ServiceRequest
-from prr.response import ServiceResponse
+from prr.utils.config import load_config
+from prr.runner.request import ServiceRequest
+from prr.runner.response import ServiceResponse
 
 config = load_config()
