+++ prr/services/providers/openai/chat.py
@@ -1,8 +1,8 @@
 import openai
 
-from prr.config import load_config
-from prr.request import ServiceRequest
-from prr.response import ServiceResponse
+from prr.utils.config import load_config
+from prr.runner.request import ServiceRequest
+from prr.runner.response import ServiceResponse
 
 config = load_config()
 openai.api_key = config["OPENAI_API_KEY"]
