+++ prr/services/service_registry.py
