+++ test/prompt/test_prompt_config.py
@@ -1,9 +1,9 @@
 import os
 import sys
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..', '..')))
 
 from prr.prompt.prompt_config import PromptConfig
-from prr.options import ModelOptions
+from prr.prompt.model_options import ModelOptions
 
 class TestPromptConfig:
 
@@ -45,7 +45,31 @@ services:
       assert config.option_for_service(service_name, 'temperature') == 0.42
       assert config.option_for_service(service_name, 'max_tokens') == 1337
 
-  def test_basic_services_model_list_no_options(self):
+  def test_basic_services_model_list_with_some_options(self):
+    config = PromptConfig()
+    config.load_from_config_contents("""
+prompt:
+  content: 'foo bar'
+services:
+  models:
+    - 'openai/chat/gpt-4'
+    - 'anthropic/complete/claude-v1.3'
+  options:
+    top_k: -1.337
+""")
+
+    assert config is not None
+    services = config.configured_services()
+
+    assert services == ['openai/chat/gpt-4', 'anthropic/complete/claude-v1.3']
+    
+    for service_name in services:
+      assert config.option_for_service(service_name, 'temperature') == ModelOptions.DEFAULT_OPTIONS['temperature']
+      assert config.option_for_service(service_name, 'max_tokens') == ModelOptions.DEFAULT_OPTIONS['max_tokens']
+      assert config.option_for_service(service_name, 'top_k') == -1.337
+      assert config.option_for_service(service_name, 'top_p') == ModelOptions.DEFAULT_OPTIONS['top_p']
+
+  def test_basic_services_model_list_with_no_options(self):
     config = PromptConfig()
     config.load_from_config_contents("""
 prompt:
